index.tsx etc(about diary component) complete
