+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@firebase/database-types": "^0.9.13",
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-regular-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
@@ -59,6 +60,7 @@
     "@types/node": "^18.7.16",
     "@types/react": "18",
     "@types/react-dom": "18",
+    "ts-loader": "^9.3.1",
     "typescript": "^4.8.3",
     "typescript-plugin-css-modules": "^3.4.0"
   }
