+++ src/components/diaryList/diaryList.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import styles from './diaryList.module.css';
 import DiaryCard from '../diaryCard/diaryCard';
-import { Diarys } from '../../types/types';
+import { FetchData } from '../../types/types';
 
-const DiaryList = ({myDiary, diaryList} : Diarys) => {
+const DiaryList = ({myDiary , diaryList} : { myDiary : boolean, diaryList : FetchData }) => {
     return (
         <ul className={styles.diaryList}>
-            {diaryList?.map(diary  =>{
+            {diaryList?.map(diary =>{
                 return <DiaryCard myDiary={myDiary} key={diary.id} diary={diary} />
             })}
         </ul>
